Add GetCommit
